requirements.txt
streamlit>=1.33.0
Pillow>=10.3.0
